✨🐛 #183: Fix Anthropic API key header and start counting its token usage (#184)
- Fixed the header where Anthropic API key is passed
- Started propagating token usage of Anthropic requests
- Corrected the TokenUsage interface by changing count field to integers from floats

+++ pkg/api/schemas/chat.go
@@ -43,9 +43,9 @@ type ModelResponse struct {
 }
 
 type TokenUsage struct {
-	PromptTokens   float64 `json:"promptTokens"`
-	ResponseTokens float64 `json:"responseTokens"`
-	TotalTokens    float64 `json:"totalTokens"`
+	PromptTokens   int `json:"promptTokens"`
+	ResponseTokens int `json:"responseTokens"`
+	TotalTokens    int `json:"totalTokens"`
 }
 
 // ChatMessage is a message in a chat request.
